register exam

+++ Backend/src/main/java/com/quiz/app/creditClass/CreditClassRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface CreditClassRepository extends CrudRepository<CreditClass, Integer> {
     @Query(value = "SELECT * FROM loptinchi as ltc WHERE ltc.nienkhoa = :schoolYear AND ltc.hocky" +
             " = :semester AND ltc.mamh = :subjectId AND ltc.nhom = :group", nativeQuery = true)
-    public CreditClass findByUniqueKey(String schoolYear, int semester, String subjectId,
+    public CreditClass findByUniqueKey(String schoolYear, double semester, String subjectId,
                                        int group);
 
     @Query(value = "SELECT ltc.* FROM loptinchi ltc join dangky dk on dk.maltc = ltc.maltc WHERE huylop = false group by ltc.maltc", nativeQuery = true)
